// src/parser/types/enums.ts
export enum VALUE_FORMAT_TYPES {
  BOOLEAN = 'BOOLEAN',
  STRING = 'STRING',
  NUMBER = 'NUMBER',
  NUMBER_STRING = 'NUMBER_STRING',
  NUMBER_LOCALE_STRING = 'NUMBER_LOCALE_STRING',
  DATE = 'DATE',
  DATE_STRING = 'DATE_STRING',
  DATE_ISO_STRING = 'DATE_ISO_STRING',
  DATE_LOCALE_STRING = 'DATE_LOCALE_STRING',
  DATE_HUMANIZED_STRING = 'DATE_HUMANIZED_STRING',
  DATE_COMPETENCE = 'DATE_COMPETENCE',
  DATE_COMPETENCE_STRING = 'DATE_COMPETENCE_STRING',
  CURRENCY_STRING = 'CURRENCY_STRING',
  PERCENT = 'PERCENT',
  PERCENT_STRING = 'PERCENT_STRING'
}

export enum VALUE_MERGE_STRATEGY {
  SUM = 'SUM',
  PRODUCT = 'PRODUCT',
  MIN = 'MIN',
  MAX = 'MAX'
}

export enum DIRECTION {
  RIGHT = 'RIGHT',
  DOWN = 'DOWN'
}